Id changed to integer

+++ src/main/java/com/currency/exchange/model/Currency.java
@@ -6,10 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
+import javax.xml.bind.annotation.*;
 
 @Data
 @Builder
@@ -23,14 +20,19 @@ public class Currency
 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @XmlElement(name="CcyMnrUnts", namespace = "http://www.lb.lt/WebServices/FxRates")
     private String ccyMnrUnts;
+
     @XmlElement(name="CcyNm", namespace = "http://www.lb.lt/WebServices/FxRates")
     private String ccyNm;
+
     @XmlElement(name="Ccy", namespace = "http://www.lb.lt/WebServices/FxRates")
     private String ccy;
+
     @XmlElement(name="CcyNbr", namespace = "http://www.lb.lt/WebServices/FxRates")
     private String ccyNbr;
+
+
 }
